+++ src/main/kotlin/org/wfanet/panelmatch/client/exchangetasks/ExchangeTaskMapper.kt
@@ -21,6 +21,7 @@ import org.wfanet.panelmatch.client.common.ExchangeContext
 class ExchangeTaskMapper(
   private val commutativeEncryptionTasks: CommutativeEncryptionTasks,
   private val mapReduceTasks: MapReduceTasks,
+  private val generateKeysTasks: GenerateKeysTasks,
   private val joinKeyTasks: JoinKeyTasks,
   private val privateStorageTasks: PrivateStorageTasks,
   private val sharedStorageTasks: SharedStorageTasks
@@ -35,13 +36,13 @@ class ExchangeTaskMapper(
       StepCase.INPUT_STEP -> privateStorageTasks.input(context)
       StepCase.COPY_FROM_PREVIOUS_EXCHANGE_STEP ->
         privateStorageTasks.copyFromPreviousExchange(context)
-      StepCase.INTERSECT_AND_VALIDATE_STEP -> joinKeyTasks.intersectAndValidate(context)
       StepCase.GENERATE_COMMUTATIVE_DETERMINISTIC_KEY_STEP ->
-        joinKeyTasks.generateSymmetricKey(context)
+        generateKeysTasks.generateSymmetricKey(context)
       StepCase.GENERATE_SERIALIZED_RLWE_KEYS_STEP ->
-        joinKeyTasks.generateSerializedRlweKeys(context)
-      StepCase.GENERATE_CERTIFICATE_STEP -> joinKeyTasks.generateExchangeCertificate(context)
+        generateKeysTasks.generateSerializedRlweKeys(context)
+      StepCase.GENERATE_CERTIFICATE_STEP -> generateKeysTasks.generateExchangeCertificate(context)
       StepCase.GENERATE_LOOKUP_KEYS -> joinKeyTasks.generateLookupKeys(context)
+      StepCase.INTERSECT_AND_VALIDATE_STEP -> joinKeyTasks.intersectAndValidate(context)
       StepCase.EXECUTE_PRIVATE_MEMBERSHIP_QUERIES_STEP ->
         mapReduceTasks.executePrivateMembershipQueries(context)
       StepCase.BUILD_PRIVATE_MEMBERSHIP_QUERIES_STEP ->
