+++ src/main/kotlin/org/wfanet/panelmatch/client/exchangetasks/testing/FakeJoinKeyTasks.kt
@@ -19,11 +19,6 @@ import org.wfanet.panelmatch.client.exchangetasks.JoinKeyTasks
 
 class FakeJoinKeyTasks : JoinKeyTasks {
   override fun generateLookupKeys(context: ExchangeContext) = FakeExchangeTask("lookup-key")
-  override fun generateSymmetricKey(context: ExchangeContext) = FakeExchangeTask("symmetric-key")
-  override fun generateSerializedRlweKeys(context: ExchangeContext) =
-    FakeExchangeTask("serialized-rlwe-keys")
-  override fun generateExchangeCertificate(context: ExchangeContext) =
-    FakeExchangeTask("exchange-certificate")
   override fun intersectAndValidate(context: ExchangeContext) =
     FakeExchangeTask("intersect-and-validate")
 }
